Replace azureFunction with logggerFunction project
Introduce a new `logggerFunction` project to replace the
`azureFunction` project. The `logggerFunction` is created
using the new `CreateFunctionProject` method, which simplifies
the setup of Azure Functions projects. It uses `mongoDb` as a
dependency and is referenced by the `beaches`, `culture`,
`geo`, and `volcano` microservices.

Added a `Function` constant to the `Microservices` class to
represent the `logggerFunction`. Updated `ProjectExtensions.cs`
to include the `Aspire.Hosting` namespace and added the
`CreateFunctionProject` method to encapsulate the logic for
setting up Azure Functions projects.